Seperating build and cli
