+++ lerna.json
@@ -1,4 +1,7 @@
 {
-  "lerna": "2.0.0-beta.30",
-  "version": "0.0.8"
+  "lerna": "2.0.0-beta.37",
+  "version": "0.0.8",
+  "packages": [
+    "packages/*"
+  ]
 }
