+++ packages/sw-build/package.json
@@ -1,7 +1,6 @@
 {
-  "name": "sw-cli",
+  "name": "sw-build",
   "version": "0.0.8",
-  "private": true,
   "description": "A CLI tool to generate a service worker and a file manifest making use of the sw-lib module.",
   "keywords": [
     "sw-fw",
@@ -9,11 +8,9 @@
     "caching",
     "fetch requests",
     "offline",
-    "cli"
+    "file manifest"
   ],
-  "bin": {
-    "sw-cli": "build/bin.js"
-  },
+  "main": "build/index.js",
   "files": [
     "src",
     "build"
@@ -25,15 +22,9 @@
   "license": "Apache-2.0",
   "repository": "googlechrome/sw-helpers",
   "bugs": "https://github.com/googlechrome/sw-helpers/issues",
-  "homepage": "https://github.com/GoogleChrome/sw-helpers/tree/master/packages/sw-cli",
+  "homepage": "https://github.com/GoogleChrome/sw-helpers/tree/master/packages/sw-build",
   "dependencies": {
-    "chalk": "^1.1.3",
-    "inquirer": "^2.0.0",
-    "lodash.template": "^4.4.0",
-    "minimist": "^1.2.0",
-    "mkdirp": "^0.5.1",
-    "sw-build": "0.0.8",
-    "update-notifier": "^1.0.3"
+    "sw-lib": "^0.0.8"
   },
   "devDependencies": {
     "fs-extra": "^2.0.0"
