+++ packages/sw-build/src/lib/constants.js
@@ -1,0 +1,4 @@
+module.exports = {
+  // Max File Size: 2MB
+  maximumFileSize: 2 * 1024 * 1024,
+};
