+++ packages/sw-build/src/lib/generate-file-manifest.js
@@ -4,7 +4,7 @@ const errors = require('./errors');
 
 /**
  * @example <caption>Generate a service worker for a project.</caption>
- * const swCLI = require('sw-cli');
+ * const swCLI = require('sw-build');
  *
  * swCLI.generateFileManifest({
  *   dest: './build/manifest.js'
@@ -31,7 +31,7 @@ const errors = require('./errors');
  * @return {Promise} Resolves once the service worker has been generated
  * with a precache list.
  *
- * @memberof module:sw-cli
+ * @memberof module:sw-build
  */
 const generateFileManifest = (input) => {
   if (!input || typeof input !== 'object' || input instanceof Array) {
