+++ packages/sw-build/src/lib/get-file-manifest-entries.js
@@ -7,7 +7,7 @@ const errors = require('./errors');
  * @property {String} url The URL to the asset in the manifest.
  * @property {String} revision The revision details for the file. This is a
  * hash generated by node based on the file contents.
- * @memberof module:sw-cli
+ * @memberof module:sw-build
  */
 
 
@@ -23,7 +23,7 @@ const errors = require('./errors');
  * @param {String} input.rootDirectory The directory run the glob patterns over.
  * @return {Array<ManifestEntry>} An array of ManifestEntries will include
  * a url and revision details for each file found.
- * @memberof module:sw-cli
+ * @memberof module:sw-build
  */
 const getFileManifestEntries = (input) => {
   if (!input || typeof input !== 'object' || input instanceof Array) {
