+++ packages/sw-build/src/lib/log-helper.js
@@ -1,0 +1,65 @@
+/**
+ * Copyright 2016 Google Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+**/
+'use strict';
+
+/* eslint-disable no-console */
+
+const chalk = require('chalk');
+
+/**
+ * Log helper is just a wrapper around the console to print slightly
+ * nicer / colored messages and could be extended to filter based on log
+ * level.
+ * @private
+ */
+class LogHelper {
+  /**
+   * Print a warning message to the console (Colored yellow).
+   * @param {string} msg Message to print to the console
+   * @param {Object} [runtimeInfo] Any information available at runtime that
+   * should be printed to the console.
+   */
+  warn(msg, runtimeInfo) {
+    let loggedMsg = chalk.yellow(msg);
+    if (runtimeInfo) {
+      loggedMsg += ` ${runtimeInfo}`;
+    }
+    console.warn(loggedMsg);
+  }
+
+  /**
+   * Print an error message to the console (Colored red).
+   * @param {string} msg Message to print to the console
+   * @param {Error} [err] An optional error to print to the console.
+   */
+  error(msg, err) {
+    console.error(chalk.red(msg));
+
+    if (err && err.stack) {
+      console.error(chalk.red(err.stack));
+    }
+  }
+
+  /**
+   * Print an info message to the console (Colored dim).
+   * @param {string} msg Message to print to the console
+   */
+  info(msg) {
+    console.log(chalk.dim(msg));
+  }
+}
+
+module.exports = new LogHelper();
