+++ packages/sw-build/src/lib/templates/file-manifest.js.tmpl
