+++ packages/sw-build/src/lib/templates/sw.js.tmpl
@@ -13,7 +13,7 @@ importScripts('<%= swlibPath %>');
  * By changing the file manifest manually, your users may end up not receiving
  * new versions of files because the revision hasn't changed.
  *
- * Please use sw-cli or some other tool / approach to generate the file
+ * Please use sw-build or some other tool / approach to generate the file
  * manifest which accounts for changes to local files and update the revision
  * accordingly.
  */
