+++ packages/sw-build/src/lib/utils/copy-sw-lib.js
