+++ packages/sw-build/src/lib/utils/filter-files.js
