+++ packages/sw-build/src/lib/utils/get-file-details.js
