+++ packages/sw-build/src/lib/utils/get-file-hash.js
