+++ packages/sw-build/src/lib/utils/get-file-size.js
