+++ packages/sw-build/src/lib/utils/write-file-manifest.js
