+++ packages/sw-build/test/.eslintrc
@@ -1,0 +1,10 @@
+{
+  "env": {
+    "mocha": true
+  },
+  "rules": {
+    "no-console": 0,
+    "max-len": 0,
+    "no-invalid-this": 0,
+  }
+}
