+++ packages/sw-build/test/generate-file-manifest.js
