+++ packages/sw-build/test/generate-sw.js
