+++ packages/sw-build/test/get-file-manifest-entries.js
