+++ packages/sw-build/test/utils-get-file-hash.js
