+++ packages/sw-build/test/utils-get-file-size.js
