+++ packages/sw-cli/src/cli-help.txt
