+++ packages/sw-cli/src/lib/constants.js
@@ -2,7 +2,4 @@ module.exports = {
   blacklistDirectoryNames: [
     'node_modules',
   ],
-
-  // Max File Size: 2MB
-  maximumFileSize: 2 * 1024 * 1024,
 };
