+++ packages/sw-cli/src/lib/errors.js
@@ -1,53 +1,5 @@
 module.exports = {
-  'unable-to-get-rootdir': 'Unable to get the root directory of your web app.',
-  'unable-to-get-file-extensions': 'Unable to get file extensions to ' +
-    'determine files to cache.',
-  'no-extension': 'Unable to detect a usable extension for a file in your ' +
-    'web app directory.',
-  'no-file-extensions-found': 'No files could be found that are suitable for ' +
-    'caching.',
-  'no-file-extensions-selected': 'No file extensions were selected so nothing '+
-    'would be cached.',
-  'invalid-file-manifest-name': 'The File Manifest Name must have at lease ' +
-    'one character.',
-  'unable-to-get-file-manifest-name': 'Unable to get a file manifest name.',
-  'invalid-sw-name': 'The service worker file name must have at least one' +
-    'character.',
-  'unable-to-get-sw-name': 'Unable to get a service worker file name.',
-  'unable-to-get-save-config': 'An error occured when asking to save details' +
-    'in a config file.',
-  'unable-to-get-file-hash': 'An error occured when attempting to create a ' +
-    'file hash.',
-  'unable-to-get-file-size': 'An error occured when attempting to get a ' +
-    'file size.',
-  'unable-to-glob-files': 'An error occured when globbing for files.',
-  'unable-to-make-manifest-directory': 'Unable to make output directory for ' +
-    'file manifest.',
-  'read-manifest-template-failure': 'Unable to read template for file manifest',
-  'populating-manifest-tmpl-failed': 'An error occured when populating the ' +
-    'file manifest template.',
-  'manifest-file-write-failure': 'Unable to write the file manifest.',
-  'unable-to-make-sw-directory': 'Unable to make the directories to output ' +
-    'the service worker path.',
-  'read-sw-template-failure': 'Unable to read the service worker template ' +
-    'file.',
-  'sw-write-failure': 'Unable to write the service worker file.',
-  'unable-to-copy-sw-lib': 'SW-Lib is needed by the service worker and could ' +
-    'not be copied over to your new site.',
   'no-file-extensions-to-cache': 'The file extensions to cache must be ' +
     'an array with at least one file extension without the dot at the ' +
     'start of the extension name, i.e. [\'html\']',
-  'invalid-generate-sw-input': 'The input to generateSW() must be an ' +
-    'object',
-  'invalid-root-directory': 'The supplied rootDirectory must be ' +
-    'a path as a string.',
-  'invalid-exclude-files': 'The excluded files should be an array of strings.',
-  'invalid-get-manifest-entries-input': 'The input to ' +
-    'getFileManifestEntries() must be an object.',
-  'invalid-manifest-path': 'The supplied manifest path is not a string with ' +
-    'at least one character.',
-  'invalid-manifest-entries': 'The manifest entries must be an array of ' +
-    'strings or JavaScript objects containing a url parameter.',
-  'invalid-generate-file-manifest-arg': 'The input to generateFileManifest() ' +
-    'must be an Object.',
 };
