+++ packages/sw-cli/test/cli-file-extensions.js
@@ -8,7 +8,7 @@ require('chai').should();
 
 describe('Ask for File Extensions to Cache', function() {
   const globalStubs = [];
-  const CLI_PATH = '../build/cli/index.js';
+  const CLI_PATH = '../build/index.js';
 
   afterEach(function() {
     cliHelper.endLogCapture();
