+++ packages/sw-cli/test/cli-root-dir.js
@@ -8,7 +8,7 @@ require('chai').should();
 
 describe('Ask for Root Directory', function() {
   const globalStubs = [];
-  const CLI_PATH = '../build/cli/index.js';
+  const CLI_PATH = '../build/index.js';
 
   afterEach(function() {
     cliHelper.endLogCapture();
