+++ templates/README.hbs
@@ -1,5 +1,6 @@
 <!-- To make changes, edit templates/README.hbs, not README.md! -->
-[![Build Status][travis-image]][travis-url]
+[![Travis Build Status][travis-image]][travis-url]
+[![AppVeyor Build status][appveyor-image]][appveyor-url]
 [![Dependency Status][dependency-image]][dependency-url]
 [![Dev Dependency Status][dev-dependency-image]][dev-dependency-url]
 
@@ -45,6 +46,8 @@ limitations under the License.
 [npm-image]: https://badge.fury.io/js/sw-helpers.svg
 [travis-url]: https://travis-ci.org/GoogleChrome/sw-helpers
 [travis-image]: https://travis-ci.org/GoogleChrome/sw-helpers.svg?branch=master
+[appveyor-image]: https://ci.appveyor.com/api/projects/status/4ct8ph4d34c5ifnw?svg=true
+[appvayor-url]: https://ci.appveyor.com/project/gauntface/sw-helpers
 [dependency-url]: https://david-dm.org/GoogleChrome/sw-helpers/
 [dependency-image]: https://david-dm.org/GoogleChrome/sw-helpers/status.svg
 [dev-dependency-url]: https://david-dm.org/GoogleChrome/sw-helpers?type=dev
